Lines 38 through 46 unnecessary?
Hey there, I'm very new to Python and taking it as a part of my GIS coursework, so apologies if this a stupid question. One of our assignments is to find some code, read it, and maybe make some suggestions if you notice something. 

I was wondering if lines 38 through 46 are necessary, as they would be handled by lines 30 and 31 anyway?